Read requiresDateOfBirth from world on register and remove all its  venue references
