--- src/components/organisms/DateOfBirthField/index.ts
@@ -1,1 +1,0 @@
-export { DateOfBirthField } from "./DateOfBirthField";
