+++ src/settings/worldDefaultSettings.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_ATTENDEES_TITLE = "attendees";
+export const DEFAULT_CHAT_TITLE = "Space";
+export const DEFAULT_REQUIRES_DOB = false;
